Make requested PR changes.
