+++ services/app-api/src/postgres/contractAndRates/submitContract.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from '@prisma/client'
-import { Contract } from './contractType'
+import { Contract } from '../../domain-models/contractAndRates/contractType'
 import { findContractWithHistory } from './findContractWithHistory'
 
 // Update the given revision
