+++ services/app-api/src/postgres/contractAndRates/unlockRate.ts
@@ -1,6 +1,6 @@
 import { PrismaClient } from '@prisma/client'
 import { findRateWithHistory } from './findRateWithHistory'
-import { Rate } from './rateType'
+import { Rate } from '../../domain-models/contractAndRates/rateType'
 
 // Unlock the given rate
 // * copy form data
