+++ services/app-api/src/postgres/prismaTypes.ts
@@ -19,37 +19,6 @@ type PrismaTransactionType = Omit<
     '$connect' | '$disconnect' | '$on' | '$transaction' | '$use'
 >
 
-// For use in prisma queries, include the updater in an updateInfo
-const updateInfoIncludeUpdater = {
-    include: {
-        updatedBy: true,
-    },
-}
-
-const draftContractRevisionsWithDraftRates = {
-    stateContacts: true,
-    addtlActuaryContacts: true,
-    contractDocuments: true,
-    supportingDocuments: true,
-    draftRates: {
-        include: {
-            revisions: {
-                include: {
-                    submitInfo: updateInfoIncludeUpdater,
-                    unlockInfo: updateInfoIncludeUpdater,
-                },
-                where: {
-                    submitInfoID: { not: null },
-                },
-                take: 1,
-                orderBy: {
-                    createdAt: 'desc',
-                },
-            },
-        },
-    },
-} as const
-
 type ContractRevisionTableWithRelations = ContractRevisionTable & {
     submitInfo?: UpdateInfoTableWithUpdater | null
     unlockInfo?: UpdateInfoTableWithUpdater | null
@@ -71,5 +40,3 @@ export type {
     UpdateInfoTableWithUpdater,
     ContractRevisionTableWithRelations,
 }
-
-export { updateInfoIncludeUpdater, draftContractRevisionsWithDraftRates }
