+++ services/app-api/src/testHelpers/contractAndRates/contractHelpers.ts
@@ -1,4 +1,4 @@
-import { InsertContractArgsType } from '../postgres/contractAndRates/insertContract'
+import { InsertContractArgsType } from '../../postgres/contractAndRates/insertContract'
 
 const createDraftContractData = (
     contractArgs?: Partial<InsertContractArgsType>
